requirements.txt
flask
gunicorn
pillow
numpy
opencv-python-headless